moving all components into just the components folder
