+++ src/components/ReactTab/ReactTab.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from './ReactTab.css';
 // * Components
 import StatePropsBox from './../../components/StatePropsBox/StatePropsBox';
-import StateContainer from './../StateContainer/StateContainer.jsx';
+import StateContainer from './../../components/StateContainer/StateContainer.jsx';
 import TreeDiagram from './../../components/TreeDiagram/TreeDiagram.jsx';
 
 const tabChange = (prevProps, nextProps) => {
