+++ src/devtools.js
@@ -6,8 +6,8 @@ import getStateDiffs from './stateDiff';
 import filter from './filterDOM';
 // * Components
 import AppNav from './components/AppNav/AppNav';
-import GraphQLTab from './containers/GraphQLTab/GraphQLTab';
-import ReactTab from './containers/ReactTab/ReactTab';
+import GraphQLTab from './components/GraphQLTab/GraphQLTab';
+import ReactTab from './components/ReactTab/ReactTab';
 // * CSS
 import styles from './devtools.css';
